Moved server functions  out of config
