+++ CLAUDE.md
@@ -19,7 +19,7 @@ npx mocha test/refs.test.js   # Run a single test file
 
 Tests use **Mocha** + **Chai** (expect) + **Sinon** (stubs/spies) + **proxyquire** (dependency injection). Route tests use `proxyquire` to swap out external dependencies (Solr, filesystem, database, email) so tests run without any backend services. Shared mock factories for Express req/res/user objects are in `test/helpers.js`.
 
-Middleware functions are in `config/middleware.js` (extracted from `app.js` for testability).
+Middleware functions are in `server/middleware.js` (extracted from `app.js` for testability).
 
 Utility scripts:
 ```bash
@@ -51,7 +51,7 @@ Node.js/Express app that manages a Solr search database (references/sources) wit
 
 ### Auth Middleware Chain
 
-Four middleware functions defined in `config/middleware.js` and imported by `app.js`:
+Four middleware functions defined in `server/middleware.js` and imported by `app.js`:
 - **`isLoggedIn`** — checks `req.isAuthenticated()`, populates `req.replacements` with user context (email, name, permission flags). Redirects to `/login` if not authenticated.
 - **`superuser`** — requires `permission >= 2`, redirects to `/profile` otherwise.
 - **`flashMessageCenter`** — collects flash messages into `req.replacements` for templates.
