+++ routes/refs.js
@@ -3,8 +3,8 @@ var router = express.Router();
 var log4js = require('log4js');
 var auditLogger = log4js.getLogger("audit");
 var proxyOpts = require('../config/solr-proxy');
-var solr = require('../config/solr-client');
-var db = require('../config/database-json');
+var solr = require('../server/solr-client');
+var db = require('../server/database-json');
 var fieldDescriptions = require('../config/fieldDescriptions');
 var Campaign = require('../models/campaign');
 var client = solr.createClient({ host: proxyOpts.backend.host, port: proxyOpts.backend.port, core: "rad" });
