+++ routes/sources.js
@@ -3,7 +3,7 @@ var router = express.Router();
 var log4js = require('log4js');
 var auditLogger = log4js.getLogger("audit");
 var proxyOpts = require('../config/solr-proxy');
-var solr = require('../config/solr-client');
+var solr = require('../server/solr-client');
 var fieldDescriptions = require('../config/fieldDescriptions');
 var client = solr.createClient({ host: proxyOpts.backend.host, port: proxyOpts.backend.port, core: "source" });
 const url = require('url');
