+++ server/solr-client.js
