+++ test/sources.test.js
@@ -26,7 +26,7 @@ var fakeProxy = {
 };
 
 var sourcesRouter = proxyquire('../routes/sources', {
-    '../config/solr-client': solrStub,
+    '../server/solr-client': solrStub,
     'log4js': log4jsStub,
     '../config/solr-proxy': proxyquire('../config/solr-proxy', {
         'http-proxy': fakeProxy
